Fixes to the tests
