+++ unit-tests/shared/src/test/scala-2/scala/ReflectiveProxyTest.scala
@@ -251,7 +251,10 @@ class ReflectiveProxyTest {
 
     class A
 
-    assertTrue(objNotifyTest(new A()) == 1)
+    assertThrows(classOf[IllegalMonitorStateException], objNotifyTest(new A()))
+    synchronized{
+      assertThrows(classOf[IllegalMonitorStateException], objNotifyTest(new A()))
+    }
   }
 
   @Test def shouldWorkOnJavaLangObjectClone(): Unit = {
